// src/app/pages/reports/Static Report/app/models/supplier.model.ts
export interface Supplier {
  id: number;
  age: number;
  birthDate: string;
  name: string;
  phoneNumber: number;
  salary: number;
}

// export interface Supplier {
//   supplierID: number;
//   supplierName: string;
//   supplierEmailAddress: string;
//   contactNumber: number;
//   address: string;
// }